+++ CHANGES.md
@@ -3,6 +3,8 @@
 
 ## Unreleased
 - MongoDB: Updated to pymongo 4.9
+- DynamoDB: Change CrateDB data model to use (`pk`, `data`, `aux`) columns
+  Attention: This is a breaking change.
 
 ## 2024/09/26 v0.0.26
 - MongoDB: Configure `MongoDBCrateDBConverter` after updating to commons-codec 0.0.18